add 7 additional strains selected by Jesse Bloom (#7)
I am adding the following 7 strains. I agree none of these are critical, but we have a bit of "room" left and I think we should use to add a bit more diversity.

That will give us 101 strains total (if they all grow). We will then have 29 or 30 vaccine strains, and even if we add a few strains later this should keep us below the 140 we used last time.

+ H3N2
  - A/SouthAfrica/PET37625/2025 | J.2.2:T65K,T128I,N159S,I160M,I214T
  - A/SouthAfrica/PET37604/2025 | J.2.2:P55T,T65K,N159S,S199T,I214T
  - A/Guangxi-Daxin/1643/2025 | G.1.3.1:T10M,G78S,I140M,G275D,K276E
  - A/Guangdong-Maonan/11438/2025 | J.2.2:T65K,A304T

+ H1N1
  - A/SouthAustralia/2518902563/2025 | D.3.1:I267T,P271Q,V321I
  - A/CoteDIvoire/2007/2025 | D.3.1:I185T
  - A/Hawaii/44/2025 | D.1:D86N,S157M,R205K